created RandomWalksGenerator, BeamSearchAlgorithm and model_test notebook
