+++ cayleypy/algo/__init__.py
@@ -2,3 +2,5 @@ from .bfs_bitmask import bfs_bitmask
 from .bfs_mitm import find_path_bfs_mitm
 from .bfs_numpy import bfs_numpy
 from .find_path import find_path
+from .random_walks import RandomWalksGenerator
+from .beam_search import BeamSearchAlgorithm
